Revert "Refactor: mv interf to "dto" fold and upd imports"
This reverts commit 101719cf278af7e85a697fc818a5563988526472.
