+++ src/middlewares/validateBody/schemas/orderSchema.ts
@@ -1,5 +1,5 @@
 import { JSONSchemaType } from 'ajv';
-import { Order } from '../../../dto/Order';
+import { Order, } from '../../../types/Order';
 
 /**
  * Material schema for post method
@@ -7,10 +7,7 @@ import { Order } from '../../../dto/Order';
 export const orderSchema: JSONSchemaType<Order> = {
   type: 'object',
   properties: {
-    status: {
-      type: 'string',
-      enum: ['pending', 'inProduction', 'finished', 'canceled'],
-    },
+    status: { type: 'string', enum: ['pending', 'inProduction', 'finished', 'canceled'] },
     internalCode: { type: 'string', minLength: 0, maxLength: 255 },
     description: { type: 'string', minLength: 1, maxLength: 255 },
 
@@ -61,10 +58,7 @@ export const orderSchema: JSONSchemaType<Order> = {
 export const partialOrderSchema: JSONSchemaType<Order> = {
   type: 'object',
   properties: {
-    status: {
-      type: 'string',
-      enum: ['pending', 'inProduction', 'finished', 'canceled'],
-    },
+    status: { type: 'string', enum: ['pending', 'inProduction', 'finished', 'canceled'] },
     internalCode: { type: 'string', minLength: 0, maxLength: 255 },
     description: { type: 'string', minLength: 1, maxLength: 255 },
